+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Update the Processor to forbid overload of ServerRequest instance and Client instance by the request argument.
 - Update the Processor to check if the parameter injected is of required type of the processed argument if this type is an object
 
+###Upgraded
+- Priorized routers behavior into manager
 
 ##[0.0.5-beta2] - 2017-08-26
 ###Updated